Morgan AI 0.4/BackEnd/app/requirements.txt
# Core FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.1

# OpenAI Integration
openai==1.54.0
tiktoken==0.8.0

# Vector Database - Updated for Python 3.13
pinecone-client==5.0.1

# LangChain Ecosystem - Updated for Python 3.13
langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.7
langchain-pinecone==0.2.0

# Data Processing
numpy==2.1.3
pandas==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Caching
redis==5.2.0

# HTTP Client
httpx==0.27.2
aiohttp==3.11.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2